+++ tokenpdf/token/token.py
@@ -1,9 +1,9 @@
-from abc import ABC, abstractmethod
+
 from typing import Tuple, Dict, Any
 from tokenpdf.utils.config import merge_configs
+from tokenpdf.utils.registry import RegistryClass
 
-
-class Token(ABC):
+class Token(RegistryClass):
     """Abstract base class for tokens.
     Defines the interface for all token types.
 
@@ -16,7 +16,7 @@ class Token(ABC):
     TOKEN_REGISTRY = None
 
     @classmethod
-    @abstractmethod
+    
     def supported_types(cls) -> Dict[str, Dict[str, Any]]:
         """Returns a dictionary mapping supported types to their expected configuration values
         and default values. Example: @see SideStandToken.supported_types()
@@ -44,7 +44,7 @@ class Token(ABC):
         """
         return config
 
-    @abstractmethod
+    
     def area(self, config, resources) -> Tuple[float, float]:
         """Calculates the required area in real-world units (mm)
         Should be based on the configuration of the token instance.
@@ -82,36 +82,10 @@ class Token(ABC):
             config = self.config
         return resources[key].filters(config.get("filters", {}), resources)
     
-    # Registry new subclasses
-    def __init_subclass__(cls, **kwargs):
-        super().__init_subclass__(**kwargs)
-        if Token.TOKEN_REGISTRY is not None:
-            Token.TOKEN_REGISTRY.register(cls)
 
 
-class TokenRegistry:
-    """A class responsible for managing token types."""
-
-    def __init__(self):
-        self._token_types = {}
-
-    def register(self, cls):
-        """Registers a token class.
-
-        Args:
-          cls: The token class to register.
-
-        Returns:
-
-        """
-        supported_types = cls.supported_types()
-        for token_type, config in supported_types.items():
-            ttype = token_type.lower()
-            if ttype in self._token_types:
-                raise ValueError(f"Token type '{ttype}' already registered.")
-            self._token_types[ttype] = (cls, config)
-    
-    def make(self, config, resources) -> Tuple[Token, Dict[str, Any], Dict[str, Any]]:
+    @classmethod
+    def make(cls, config, resources) -> Tuple["Token", Dict[str, Any], Dict[str, Any]]:
         """Creates a new token instance of the specified type.
 
         Args:
@@ -123,13 +97,16 @@ class TokenRegistry:
           : A new token instance.
 
         """
-        token_type = config.get("type", "Circle").lower()
-        if token_type not in self._token_types:
+        token_type = config.get("type", "circle").lower()
+        token_types = cls.get_registry()
+        if token_type not in token_types:
             raise ValueError(f"Unsupported token type: {token_type}")
-        cls, default_config = self._token_types[token_type]
+        cls, default_config = token_types[token_type]
         merged_config = merge_configs(default_config, config)
         token = cls()
         config = token.apply_defaults(merged_config, resources)
         return token, config
-
-Token.TOKEN_REGISTRY = TokenRegistry()
+    
+    @classmethod
+    def _get_class_registry_args(cls, name) -> Dict[str, Any]:
+        return cls.supported_types().get(name, {})